pyproject.toml
[project]
name = "codenamesbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.75.0",
    "discord-py>=2.6.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
]
